msaa: Add test to verify sample-alpha-to-coverage with integer draw buffers
When draw buffer zero is bound to an integer format buffer,
GL_SAMPLE_ALPHA_TO_COVERAGE should have no effect. This should work
properly even if there are other draw buffers that are not in integer
format.

Note: Test passes on Sandybridge with following patch applied on mesa.
Patch is under review on mesa-dev mailing list:
[PATCH] i965/msaa: Disable alpha-to-{coverage, one} when drawbuffer zero
is in integer format

Reviewed-by: Paul Berry <stereotype441@gmail.com>