msaa: Test proper functioning when no color buffer is present.
This test reproduces a problem in Mesa's i965 driver as of 7/12/12
(commit 1a06e84): when no color buffer is present, but a multisampled
depth or stencil buffer is present, Mesa fails to set up the rendering
pipeline for multisampling.

The test includes variants for testing both the stencil buffer and the
depth buffer.  The depth buffer test may be run in a "computed" mode,
in which case the depth values are output by the fragment shader (this
is an important corner case for i965).

Verified on nVidia's proprietary binary driver for Linux.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>
Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>