Test that fragment shader writes to render targets are properly ordered.
Some hardware platforms (e.g. i965 gen6+) are capable of processing
the fragments corresponding to multiple primitives in parallel, even
if those primitives overlap on the screen.  This test verifies that in
such situations, the fragment shader outputs are written to the
framebuffer in the correct order.

v2: Add to all.tests.

Reviewed-by: Eric Anholt <eric@anholt.net>