Fix swapbuffers timing in arb_seamless_cubemap test.
Prior to this patch, the test would:
1. Draw into the back buffer.
2. glutSwapBuffers
3. Read/probe pixel values...from what used to be the front buffer?

This patch moves the SwapBuffers to after the probing, which should
ensure that draws and reads use the same buffer.  It also switches
from glutSwapBuffers to piglit_present_results().

This may help with bug #46212.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=46212
Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Brian Paul <brianp@vmware.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>