copyteximage: Fix the test to eliminate GL_TEXTURE_1D failures on NVIDIA
Draw and probe a QUAD (in place of a LINE) in case of GL_TEXTURE_1D. This
fixes the test failures on NVIDIA's proprietary linux drivers. In case of
LINE, glReadPixels() misses the relevant pixels by one pixel on NVIDIA.

V2: Draw taller quad to avoid the risk of hitting specific pixels in
glReadPixels.

Signed-off-by: Anuj Phogat <anuj.phogat@gmail.com>
Reviewed-by: Brian Paul <brianp@vmware.com>