util-cl: Add utilities to support OpenCL tests
piglit-util-cl.* adds functions for faster and easier writing of OpenCL
programs. Most of the functions return NULL or false on errors and print
the error to stderr, others print an error and exit the program.

piglit-util-cl-enum.* files add enum definitions (some of them also by
version) and their string representations.

Signed-off-by: Blaž Tomažič <blaz.tomazic@gmail.com>