cl: Add OpenCL tests
All of the tests are located in tests/cl. Each type of tests are located
in their own subfolders. Custom tests in custom folder, API in api and
Program in program.

Program folder has special tests that are run by program-tester.c which
parses .cl or .program_test tests. This type of tests test program
building and kernel execution. These tests are located in program/build
and program/execute.

Signed-off-by: Blaž Tomažič <blaz.tomazic@gmail.com>