Fix explicit attrib location tests layout-12|13.frag; add a new one.
The ARB_explicit_attrib_location tests layout-12.frag and layout-13.frag
were incorrect: they both used index == 2 (which is invalid) and
expected the compile to succeed.  Although the extension spec doesn't
state anything about generating an error on indices other than 0 or 1,
the GLSL 4.30 spec clarifies this as a compile error.  This matches the
behavior of the API call defined in ARB_blend_func_extended,
glBindFragDataLocationIndexed, which generates INVALID_VALUE for indices
other than 0 or 1.

This patch changes layout-13.frag to use index = 1 so it will actually
test the intended behavior.

layout-12.frag was just bogus: the comment said "Even though the
specified locations overlap [...]" but the test only had one variable.
This patch turns it into a negative test to ensure index == 2 is a
compile error.

This patch also adds layout-14.frag, a test to check that index < 0 is
also a compile error.