Test that sRGB-related blits are allowed and do not do encoding/decoding.
The behaviour specified by OpenGL for blits involving sRGB is
self-contradictory--it is unclear whether blits should perform sRGB
encoding/decoding, and if so, whether this encoding/decoding should be
dependent upon the setting of the GL_FRAMEBUFFER_SRGB enable flag.
Experiments with nVidia and ATI drivers indicate that they never do
any sRGB encoding or decoding during a blit, regardless of the buffer
formats and regardless of the setting of GL_FRAMEBUFFER_SRGB.

Furthermore, OpenGL specifies that multisampled blits between formats
that are not "identical" should fail, however nVidia and ATI drivers
allow these blits.

Existing games, such as Valve's "Left For Dead 2", appear to rely on
the behaviour exhibited by the the nVidia and ATI drivers, so it seems
sensible for Piglit to test for this behaviour too.

Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>