util: Add new GL test frameworks
This patch adds three major types of GL test frameworks, listed below.
They are described in detail in this patch's README.txt.
    piglit_glut_framework   (uses GLUT)
    piglit_winsys_framework (uses Waffle)
    piglit_fbo_framework    (uses Waffle)

This patch only adds and builds the source for new frameworks; they are
not yet used anywhere. Subsequent commits switch Piglit over to use the
new frameworks and delete the sources for the old framework code.

For details about the new frameworks, see this patch's README.txt.

Reviewed-and-tested-by: Jordan Justen <jordan.l.justen@intel.com>
Tested-by: Paul Berry <stereotype441@gmail.com>
Acked-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>