cmake: Define cmake flags and feature macros for X11, GLX, EGL
On Linux, unconditionally set the following internal CMake flags to true:
    PIGLIT_HAS_X11
    PIGLIT_HAS_GLX
Also define feature macros with the same names.

On all systems, if libEGL is found, then do the same for PIGLIT_HAS_EGL.

The CMake flags will allow us to selectively add certain files to
libpiglitutil_gl*. The flags and macros are used in a following commit.

Reviewed-and-tested-by: Jordan Justen <jordan.l.justen@intel.com>
Tested-by: Paul Berry <stereotype441@gmail.com>
Acked-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>