fbo-clear-formats: test stencil glClear
The test visualizes the stencil buffer (unlike some other tests),
which is very useful for debugging.

Unfortunately, it throws a weird GL error when run multiple times in a row,
which can be reproduced by running:

./fbo-clear-formats GL_EXT_packed_depth_stencil stencil -auto

Commenting out the last glDeleteTextures call gets around the issue.
I think it's a bug in Mesa.