piglit: fix crashes when using original GLUT
The glutInit() function in the original GLUT expects argv[0] to be the
program name (as from main()), not a NULL pointer.  A strrchr() call
crashes on NULL.

For now, set argv[0]="piglit".  Previously, glutInit would get the
program path/name and it would appear in the window title.

This seems to be a regression since the waffle changes.

Reviewed-by: José Fonseca <jfonseca@vmware.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com>