util/gl: Add test config option 'requires_displayed_window'
When running tests in -auto mode using waffle, piglit does not display the
window on the screen. The nice benefit of this is that test runs are
visually silent except for the rare glean and glx tests.

However, not displaying the window breaks tests that read from the front
buffer. After glXSwapBuffers, the content of the front buffer is undefined
if the window is not displayed.

Reviewed-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>