clearbuffer-display-lists: Fix initial clear color test.
The test claimed that the default value was 0,0,0,1, but the GL 3.0
spec page 263 disagrees.  The test saw 0,0,0,1 when it was written
because a visual with no alpha happened to be chosen, but there's no
guarantee that you don't have alpha, so fix the test, and ask for
alpha in the test so that we can reliably check those bits.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>