util/gl: add default window config of 150 x 150 pixels
On Win7 windows get resized to the min size of 116x100 pixels (which
might depend on the Aero theme, etc).  This resize is detected by
the piglit framework:

Got spurious window resize in automatic run (100,100 to 116,100)
PIGLIT: {'result': 'warn'}

so we get a 'warn' result instead of actually running the test.

This sets a default window size of 150x150 pixels.  Next, we'll
remove the "config.window_width/height = 100" lines from a bunch
of tests.

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>