time-elapsed: Spend time in the FS rather than do a lot of draws.
The time-elapsed test needs to consume a lot of GPU time so that it
can validate GPU time queries against elapsed wall-clock time.
Previously we did this by painting the window thousands of times.
That sometimes caused false failures due to the CPU time consumed in
preparing the draw calls.

This patch changes the test so that it consumes time by using a slow
fragment shader rather than by painting the window over and over.