glx-multithread: Replace glFlush calls with glFinish
Currently, in Mesa with the i965 driver, calling glFlush() forces all
pending drawing operations to be sent to the kernel; this ensures that
any subsequent drawing operations will be performed after those
pending drawing operations, even if they come from other contexts.
The glx-multithread test was relying on this behaviour, however, the
GL and GLX specs do not garantee this.

This patch modifies the glx-multithread test to use glFinish() instead
of glFlush().

Reviewed-by: Chad Versace <chad.versace@linux.intel.com>
Reviewed-by: Eric Anholt <eric@anholt.net>