cl: Remove clc_version_min requirement on scalar-load-[float|uchar].cl tests
Nothing in these tests should require CL C 1.1.

If the tests were attempting to WRITE to a global char*/short*,
then cl_khr_byte_addressable_store or CL1.1 (not CLC 1.1) would be required.

Also added a param/constant test case to the uchar file so that it's
consistent with the float tests.

Reviewed-by: Tom Stellard <thomas.stellard@amd.com>