arb_internalformat_query: Verify minimum maximums
This test passes on NVIDIA's closed-source driver (version 304.64 on
GTX260).  This driver supports GL_ARB_texture_multisample.

This test fails on AMD's closed-source driver (Catalyst 12.6 on HD
3650).  Querying GL_SRGB8_ALPHA8 (for all targets) returns {1, 0, 2}.
This is invalid because it is not descending order, and zero is not
positive.  I suspect they intended to return 1 for NUM_SAMPLE_COUNTS,
and {1} for SAMPLES.  This driver supports GL_ARB_texture_multisample.

v2: Delete the ARB_texture_multisample defines.  Recent changes to
enumsext.spec make this unnecessary.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Eric Anholt <eric@anholt.net>