cl: Fix warnings-as-errors CL test.
I had previously been under the impression that an unused variable would be
enough to trigger a warning and therefore get the test to pass.. I was wrong.

Now, we attempt to enable a made up extension name, which throws a warning and
causes the program to fail to build when the -Werror build option is present.