shader_runner: Parse ES versions
Now shader_runner accepts an ES qualifier in version requirements. For
example, the following are now allowed:
    GL >= 3.0 es
    GLSL >= 3.00 es

shader_runner is not yet capable of creating an ES context. That requires
additional massaging in follow-on patches.  Therefore, if a test requests
an ES version, shader_runner raises an assertion failure in
get_required_versions().

Signed-off-by: Chad Versace <chad.versace@linux.intel.com>