piglit-dispatch: Update enumext.spec from the registry
The most notable changes are the removal of GL_TEXTURE_BUFFER_FORMAT
(this was not prototed to core from the ARB extension), and the
addition of a giant pile of new extesions.

Support is added to piglit_get_gl_enum_name for
GL_SHADER_BINARY_FORMATS which was previously missing from our
enumext.spec.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>