arb_get_program_binary: Verify glGet* doesn't over-run the buffer for binary format queries
This test works with both GL_PROGRAM_BINARY_FORMATS (from
ARB_get_program_binary) and GL_SHADER_BINARY_FORMATS (from
ARB_ES2_compatibility).  Mesa fails GL_SHADER_BINARY_FORMATS because it
generates an error (!) for the queries.  Fix the error causes it to
over-run the buffer.  I suspect the over-run will happen in an ES2 with
Mesa as-is.

v2: Use less terse function and parameter names.  Avoid the use of
rand().  Both suggested by Paul.  Also move the empty piglit_display
function to the end of the file.  This matches changes in the piglit
test convention that were adopted after this test was initially written.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>