arb_get_program_binary: Execercise GL_PROGRAM_BINARY_RETRIEVABLE_HINT
This test fails on NVIDIA (304.64 on GTX 260).  It fails because the
value of GL_PROGRAM_BINARY_RETRIEVABLE_HINT is not "patched" by
glLinkProgram.  The spec is a little unclear on this topic, and I filed
a bug in the Khronos bugzilla.  After some discussion about which state
should be latches vs. live, the consensus was that
GL_PROGRAM_BINARY_RETRIEVABLE_HINT should be latched.  Updated language
has been proposed, but a new spec has not been spun.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Reviewed-by: Paul Berry <stereotype441@gmail.com>
Cc: Wen Su <WSu@nvidia.com>