piglt: Add support for subtests v2
Subtests allow a test to produce more than one result.  This is useful
for tests where it may be convenient to combine several somewhat
unrelated tests into a single program.  For example, a cl test like:

kernel void test(TYPE *out, TYPE *in) {
	out[0] = some_builtin_function(in[0], in[1]);
}

That uses only one kernel, but recompiles it several times with different
definitions of the macro 'TYPE' in order to test the builtin with all
the possible types.

To take advantage of the subtest feature, programs should output one
PIGLIT line per subtest as in the following example:

PIGLIT:subtest {'testA' : 'pass'}
PIGLIT:subtest {'testB' : 'fail'}

Where testA and testB are the name of the subtests.

In the result summary, this will be displayed as:

TestName                            1/2
      testA                         pass
      testB                         fail

v2:
  - Print one line for each subtest rather than printing them all
    together.

Acked-by: Chad Versace <chad.versace@linux.intel.com>