framework/shader_test: Assume the api is GL if no requirements are found
Commit 260f211 caused some shader tests to regress with the message:
"Failed to find [require] block".

This patch changes the python script that drives shader_runner to assume
that the test's API is GL if no GL requirement is found in the test file.
This behavior matches the behavior of the shader_runner executable, whose
default requirements are GL >= 1.0 and GLSL >= 1.10.

I tested this patch with a full Piglit run. The delta against
master-ed3903c consists only of fixes.

Reviewed-by: Tom Gall <tom.gall@linaro.org>
Tested-by: Tom Gall <tom.gall@linaro.org>
Reported-by: Ian Romanick <idr@freedesktop.org>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>