all.tests: fix the mess with VAO tests
Please people, when adding new tests, test whole piglit and compare
the results with origin/master, so that you can be sure you don't remove
any existing tests and your new tests are executed as expected.

This commit adds back the existing VAO tests and fixes the filename
of the new isvertexarray test.