shader_runner_gles2: Remove unused variable warnings
The function macros in shader_runner_gles_workarounds.h cause many unused
variable warnings. The function macros have form
  #define glMapBuffer(...) UNDEFINED_FUNCTION(glMapBuffer, 0)
and get expanded to the call
  unsupported_function("glMapBuffer", 0)
Observe that the original arguments to glMapBuffer are unused.

This patch fixes the warnings by changing the macro form to
  #define glMapBuffer(...) UNDEFINED_FUNCTION(glMapBuffer, 0, __VA_ARGS__)
which gets expanded to the call
  unsupported_function("glMapBuffer", 0, __VA_ARGS__)

Reviewed-by: Tom Gall <tom.gall@linaro.org>
Tested-by: Tom Gall <tom.gall@linaro.org>
Signed-off-by: Chad Versace <chad.versace@linux.intel.com>