Remove commen-continuation test.
This test was originally designed to verify that a comment line ending
with a backslash would not be considered a line-continuation
character, (since a shader has been seen in the wild with such a
backslash and where interpreting the backslash as a line-continuation
would cause compilation failure).

However, with OpenGL 4.2 and GLES3, backslash as a line continuation
character is now specified to operate globally. So this test is now
invalid, and the correct result is that this shader should cause a
compilation failure.

[Note that Mesa does have support for a quirk that can disable the
line continuation on an app-by-app basis with driconf
configuation. But that's something that piglit itself can't easily
test. Instead, Mesa now includes its own testing for that.]