fbo-blending-formats: Enable testing for channels with 0 bits of data.
The existing code is intentionally disabling testing for channels with
1 bit of data, (since the ambiguity in rounding makes it infeasible to
test). But the disabling of testing for 0-bit channels was
unintentional collateral damage.

In particular, for RGB-only formats it's important to test that
results always provide a value of 1.0 for alpha.

So we change the condition to disable tests only when the number of
bits is equal to 1, (as opposed to less then 2). We also just fudge in
a tolerance value corresponding to 8 bits for channels with 0 bits of
storage. (The tolerance probably doesn't matter at all here since we
should really only be doing exact comparisons of values of 0.0 or
1.0).

Reviewed-by: Marek Olšák <maraeo@gmail.com>