fbo-blending-formats: Test DST_ALPHA blending when destination has no alpha
It may sound silly to test the blending of DST_ALPHA when the
destination has no alpha channel, but in fact this is a not uncommon
source of errors in implementations.

Some implementations will implement a non-alpha format with an
internal buffer that actually does have bits for the alpha
channel. These implementations must then ensure that these alpha bits
never leak out, and must have additional code to smash alpha to 1.0
for things like ReadPixels, TexImage, and DST_ALPHA blending.

So, fix the test to not avoid this important testing.

Reviewed-by: Marek Olšák <maraeo@gmail.com>