arb_texture_multisample: add new test for errors
Tests that glFramebufferTextureLayer produces the correct error for
layer < 0 when used with GL_TEXTURE_2D_MULTISAMPLE_ARRAY. This was
overlooked in the initial mesa implementation of
ARB_texture_multisample, and crashed deep in the driver instead.

V3: - Don't set the window size, we don't care.
V4: - Add missing copyright notice.

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>