texture-integer-glsl130: fix random failures with wrong expected alpha values
The logic failed whenever the random number was 0 for formats which
don't have an alpha channel. In this case all of expected/bias/value
were 0 but the actual correct read back value is 1.0 (default value
for alpha channel sampling for integer textures is 1 just like it is
1.0 for ordinary texture formats).

Reviewed-by: Brian Paul <brianp@vmware.com>
Reviewed-by: Jose Fonseca <jfonseca@vmware.com>

Edited-by: Carl Worth <cworth@cworth.org>

Use integer literals, not floating-point to assign to integer
array. Add a comment describing the tweak of values for texture
formats without some channels.