glx-multithread-makecurrent-*: Write the expected value for alpha
The tests would previously write 0.0 to the alpha and expect to read 1.0
back.  If the test got a visual without alpha (and it doesn't request
alpha), this would happen to work.  However, drivers are not required to
expose visuals without alpha, so the test might get an RGBA visual
afterall.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=59738
Reviewed-by: Brian Paul <brianp@vmware.com>