glsl-es-1.00: Move contents of glsl-1.00 to glsl-es-1.00
Move contents of glsl-1.00 to the glsl-es-1.00 directory.

Remove a few blank lines at the end of 4 tests so that git is happy.

Last fix up all.tests so that all shader and glslparser tests for
glsl-es-1.00 are listed and referenced in the new directory.

V2: patch squashed and generated with --find-renames

Signed-off-by: Tom Gall <tom.gall@linaro.org>

V3 (Paul Berry <stereotype441@gmail.com>): Add missing initialization
of spec['glsl-es-1.00']['execution'] to all.tests.

Reviewed-by: Paul Berry <stereotype441@gmail.com>