generated_tests: Fix geometry shader tests to use gl_VerticesIn
Commit ff315ae (Add geometry testing to code-generated tests)
erroneously omitted the parentheses when calling
glPositionIn.length().  It also failed to account for the fact that
the .length() method is unavailable in GLSL 1.10.

This patch modifies the code to use gl_VerticesIn instead, since this
is provided by ARB_geometry_shader4, and hence is available regardless
of the GLSL version being tested.

Thanks to Ken Phillis Jr for discovering this bug.

Cc: Ken Phillis Jr <kphillisjr@gmail.com>