Change expected behavior for #if with undefined macros.
The GLSL 1.30 specification deviated from the standard behavior of C
preprocessors everywhere by making an undefined macro an error rather
than treating it as a value of 0. This was not actually useful, and
this exception to the standard behavior is dropped in GLSL 4.30.

In the meantime, popular implementations have been ignoring the
specified behavior and implementing the standard behavior anyway. So
expect that.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=51631

Reviewed-by: Anuj Phogat <anuj.phogat@gmail.com>