depthstencil-render-miplevels: Present the results in non-auto mode.
I tried to make the presentation match the data that was originally
probed out of the buffer.  This means not using depth texturing, since
on Intel that can trigger texture relayouts.  I also left in use of
the piglit probe functions for the auto path, so there are two
separate readpixels calls on each level in the non-auto case where
presentation happens.

v2: Use Paul's ortho hack for large miplevels, and rewrite commit message.