glsl-1.30: add basic test for testing textureOffset functionality.
This is pretty rough and doesn't really test all that much (just
textureOffsetLod, but there's other texturing functions with offsets),
doesn't test different wrap modes, NPOT sizes etc., but there's no
other test using "ordinary" texture opcodes and texel offsets, so
it's a start (and in fact it exposed a bug in the mesa state tracker).
Tested with llvmpipe (pass) and softpipe (fail) which is as expected -
hopefully the math is ok...
Inspired from fs-texelFetchOffset.

v2: some minor comment and code fixes

Reviewed-by: Jose Fonseca <jfonseca@vmware.com>
Reviewed-by: Brian Paul <brianp@vmware.com>