fbo-maxsize: minor test improvements
Clear the large FBO to a non-black color and probe for it in the
final test.

This uncovers a bug in Mesa's llvmpipe driver where we ran out of
bin commmand memory and skipped some tiles, leaving them black.

Setting the clear color to white (and actually clearing the FBO)
helps to detect that.