arb_texture_multisample: add new negative-max-samples test
This covers the sample count limits for both renderbuffers and
multisample textures, when ARB_texture_multisample is supported.

Skips if ARB_internalformat_query is supported, since that extension
allows the overall sample count limits to be exceeded on a
per-internalformat basis.

V2: - Emit a useful message when skipping.
V3: - Style nits

Signed-off-by: Chris Forbes <chrisf@ijw.co.nz>
Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>