Add tests for the GL_AMD_performance_monitor extension.
One of the challenging aspects of testing this extension is that it
defines an implementation-specific set of groups and counters.  Many of
the tests here arbitrarily operate on the counters in the first group,
while a few sanity check all counters in all groups.

All tests pass on AMD's Catalyst 12.06 driver on a Radeon 3650, except
for: test_number_of_groups_partial_array, test_delete_monitor_invalid,
and test_get_counter_data_byte_size.  I believe the first two are
legitimate bugs in AMD's implementation, and the latter is a
disagreement about the associativity of text within a paragraph.  Still,
these are minor details and I believe the tests are valid.

v2: Actually set the test name for the second test.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>