framework: hardware independent interface for dma buffers
In order to test the EXT_image_dma_buf_import, one needs a way for
creating dma buffers that can be imported to EGL and filling them
with data for the GL-stack to sample.
While dma buffer themselves are only defined for linux, the actual
writing of the buffers using CPU differs from hardware to another,
and possibly from window system to another. The intention here is
to push these details into the framework leaving the actual tests
environment independent.

v2 (Chad):
   - replace 'void*' buffer type by 'struct piglit_dma_buf *'

v3 (Chan, Ken):
   - document passing NULL pointers to buffer desctruction

Signed-off-by: Topi Pohjolainen <topi.pohjolainen@intel.com>
Reviewed-by: Chad Versace <chad.versace@linux.intel.com> (v2)
Reviewed-by: Eric Anholt <eric@anholt.net>