arb_shader_texture_lod-texgradcube: Test explicit derivatives for cube maps
Similar to arb_shader_texture_lod-texgrad, but using cube maps instead.
Given the somewhat undefined behavior of explicit gradients with cube
maps, the main purpose of the test is really to test that those work at
all, as there doesn't seem to be any other test covering this.
That said, drivers which simply drop explicit derivatives on the floor
(like softpipe) pass this perfectly as it simply compares implicit vs.
explicit behavior (which, given the fuzzy specification, might not be
really required to be the same here, though given the chosen values, that
is the major axis derivatives being zero, it might seem like a reasonable
assumption).
I guess something which would also test that the implementation is really
using explicit derivatives instead of implicit ones would also be desirable,
but I'll leave that for now, I couldn't really come up with something.

Reviewed-by: Brian Paul <brianp@vmware.com>