Test that transform feedback works with non-flat integers.
Commit 0fac0a6 (Modify tests for error checking on "flat" keyword to
match updated specs) adopted the GLSL 1.50 rules for "flat" varyings
for all desktop GL tests (the rationale was that this change likely
reflects a bug fix rather than a deliberate behavioural difference).

The GLSL 1.50 rules allow a vertex shader output to be a non-flat
integer provided that it isn't consumed by the fragment shader.  In
particular, this means that a non-flat integer may be used with
transform feedback.

This test verifies that transform feedback works correctly for all
integral types, even if they are not declared as "flat".

Reviewed-by: Jordan Justen <jordan.l.justen@intel.com>