array-of-struct-of-array: Added indexing operator to array.
The GLSL 1.2 spec says: "Arrays can also [be] accessed with the method operator
( . ) and the length method to query the size of the array:".

It doesn't say that members of arrays of structs can be accessed with the method
operator even if the dereference is only used to access the length method of a
nested array.

Reviewed-by: Matt Turner <mattst88@gmail.com>