html: Change the main table from "width: 100%" to min-width.
When a report contains many columns, the table may need to be larger
than the width of the page.  Currently that happens due to the browser
looking at the table content in order to determine the size, but when we
switch to table-layout: fixed, it would limit it to the page size.

Doing this now avoids the problem before it first occurs.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>