summary.py: Change skip to 'Not Run' for tests not run
When comparing two test results the current behavior sets a status of
skip for tests that were not run in one set of results. This is
confusing and silly, so instead give them a status of N/A which gives a
clear understanding of what skipped, and what simply didn't run either
by exclusion or because it didn't exist yet.

V3: - Change N/A to "Not Run"
V4: - Fixes "Not Run" to not be compared for fixes, regressions or
	  problems

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Reviewed-By: Aaron Watry <awatry@gmail.com>
Reviewed-By: Kenneth Graunke <kenneth@whitecape.org>