piglit: Prefer waffle's getprocaddress/dlsym() support to our own.
Notably, waffle has EGL bits which this stuff doesn't.  We can't
globally enable this and remove the old platform-specific
dlsym/getprocaddress code, because waffle isn't always built against
yet, and even when we're building against it, it's not used in some of
the EGL/GLX tests (yet, at least).

v2: Note that the non-waffle functions are ignored in the waffle
    build.

Reviewed-by: Paul Berry <stereotype441@gmail.com>