piglit: Make callers of piglit_dispatch_default_init() specify the API.
To automatically detect the current API, we'd need to make calls to
piglit_is_gles() and piglit_get_gl_version(), while bypassing the
dispatch mechanism we're trying to initialize.  That seems tricky,
while the context type is probably going to be known by the caller
when we extend dispatch to GLES tests.

Reviewed-by: Paul Berry <stereotype441@gmail.com>